Implemented table storage delete
Removed Id and use per-entity key, removed CreatedOn because we can use Timestamp,

+++ Tests/sfa.Tl.Marketing.Communication.IntegrationTests/ProviderUrlTests.cs
@@ -115,9 +115,9 @@ namespace sfa.Tl.Marketing.Communication.IntegrationTests
 
             var cloudTableClient = cloudStorageAccount.CreateCloudTableClient();
 
-            var providerRepository = new GenericCloudTableRepository<ProviderEntity, int>(
+            var providerRepository = new GenericCloudTableRepository<ProviderEntity, long>(
                 cloudTableClient,
-                loggerFactory.CreateLogger<GenericCloudTableRepository<ProviderEntity, int>>());
+                loggerFactory.CreateLogger<GenericCloudTableRepository<ProviderEntity, long>>());
 
             var qualificationRepository =  new GenericCloudTableRepository<QualificationEntity, int>(
                     cloudTableClient,
